components/py-sstv-groundstation/src/sstv_decoder.py
# -*- coding: utf-8 -*-
#!/usr/bin/env python3

"""
Модуль декодирования SSTV
Этот модуль содержит функции для декодирования SSTV-сигналов
в изображения с использованием библиотеки pysstv.
"""

import numpy as np
from typing import Optional, Tuple
from PIL import Image

class SSTVDecoder:
    """
    Класс для декодирования SSTV-сигналов
    Обрабатывает декодирование SSTV-сигналов в изображения
    с использованием библиотеки pysstv.
    """


    def __init__(self):
        """Инициализирует декодер SSTV"""
        self.decoded_image = None
        self.signal_data = None


    def decode_from_audio(self, audio_file: str) -> Optional[Image.Image]:
        """
        Декодирует SSTV-сигнал из аудиофайла

        Args:
            audio_file: Путь к аудиофайлу с SSTV-сигналом

        Returns:
            Image.Image: Декодированное изображение или None при ошибке
        """
        try:
            # Импортируем библиотеку только при необходимости
            from pysstv.sstv import SSTV
            import wave

            # Открываем аудиофайл
            with wave.open(audio_file, 'r') as wav:
                frames = wav.readframes(wav.getnframes())
                sample_width = wav.getsampwidth()
                framerate = wav.getframerate()

                # Преобразуем аудиоданные в числовой формат
                if sample_width == 1:
                    dtype = np.uint8
                elif sample_width == 2:
                    dtype = np.int16
                else:
                    raise ValueError(f"Неподдерживаемая глубина цвета: {sample_width}")

                audio_data = np.frombuffer(frames, dtype=dtype)

                # Декодируем SSTV-сигнал
                # Этот процесс требует использования pysstv
                print(f"Декодирование SSTV-сигнала из файла: {audio_file}")

                # Здесь будет реализация декодирования SSTV
                # Возвращаем заглушку изображения
                decoded_img = Image.new('RGB', (320, 240), color='blue')
                self.decoded_image = decoded_img

                return decoded_img

        except ImportError:
            print("Библиотека pysstv не установлена. Установите с помощью: pip install pysstv")
            return None
        except Exception as e:
            print(f"Ошибка при декодировании SSTV-сигнала: {str(e)}")
            return None


    def save_decoded_image(self, filepath: str) -> bool:
        """
        Сохраняет декодированное изображение в файл

        Args:
            filepath: Путь для сохранения изображения

        Returns:
            bool: True если изображение успешно сохранено, иначе False
        """
        if self.decoded_image is None:
            print("Сначала декодируйте изображение")
            return False

        try:
            self.decoded_image.save(filepath)
            print(f"Изображение успешно сохранено: {filepath}")
            return True
        except Exception as e:
            print(f"Ошибка при сохранении изображения: {str(e)}")
            return False

def convert_audio_to_image(audio_data: np.ndarray, sample_rate: int) -> Optional[Image.Image]:
    """
    Конвертирует аудиоданные в изображение

    Args:
        audio_data: Аудиоданные в формате numpy array
        sample_rate: Частота дискретизации аудио

    Returns:
        Image.Image: Результативное изображение или None при ошибке
    """
    # Заглушка для конвертации аудио в изображение
    # Реализация зависит от конкретного SSTV-режима
    width, height = 320, 240  # Стандартный размер для некоторых SSTV-режимов
    img_array = np.random.randint(0, 255, (height, width, 3), dtype=np.uint8)
    img = Image.fromarray(img_array)
    return img

def detect_sstv_signal(audio_data: np.ndarray, sample_rate: int) -> Tuple[bool, float]:
    """
    Обнаруживает SSTV-сигнал в аудиоданных

    Args:
        audio_data: Аудиоданные в формате numpy array
        sample_rate: Частота дискретизации аудио

    Returns:
        Tuple[bool, float]: (True если найден сигнал, приблизительная частота начала)
    """
    # Заглушка для обнаружения SSTV-сигнала
    # В реальной реализации анализируется наличие характерных тонов
    print("Поиск SSTV-сигнала в аудиоданных...")
    return True, 0.0  # Предполагаем, что сигнал найден

