requirements.txt
# Python requirements for Nanoprobe Simulation Lab
# Core dependencies for the main project

# For image processing and analysis
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0
Pillow>=8.3.0
PyYAML>=6.0
scipy>=1.7.0
scikit-learn>=1.0.0
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.30.0
psutil>=5.8.0

# For validation
jsonschema>=4.0.0

# For SSTV decoding (rtlsdr is optional, only needed if you have RTL-SDR hardware)
pysstv>=0.2.0
#rtlsdr>=0.2.0

# For development and testing
pytest>=6.0.0
memory-profiler>=0.60.0
objgraph>=3.5.0

# For scheduling and automation
schedule>=1.2.0

# For visualization and dashboards
plotly>=5.0.0
seaborn>=0.11.0

# Additional dependencies
requests>=2.28.0